refactor: extract most of tooltip logic into reusable mixin
