+++ packages/tooltip/test/tooltip-timers.test.js
@@ -11,7 +11,8 @@ import {
 } from '@vaadin/testing-helpers';
 import sinon from 'sinon';
 import './not-animated-styles.js';
-import { resetGlobalTooltipState, Tooltip } from '../vaadin-tooltip.js';
+import { resetGlobalTooltipState } from '../src/vaadin-tooltip-mixin.js';
+import { Tooltip } from '../vaadin-tooltip.js';
 import { mouseenter, mouseleave } from './helpers.js';
 
 describe('timers', () => {
